refactored the audio feedback into a separate class
